Make compatible with updated amrex/c++20
This commit makes the following changes, intended to
make the example compatible with the most recent
amrex version, which requires c++20:
1. The requested compiler is switched to c++20 instead
of c++17.
2. All ParallelFor loops are given the this pointer
in the capture statement of the lambda function (this
only addresses a warning associated with leaving
out the this pointer when compiling with c++20),
3. in StateVariables, the subtraction between NUM_VARS
and NUM_CCZ4_VARS in the addition of phi and pi is
done by explicitly casting these enums to integers,
for the same reason as #2 above.

This code compiles but has not been run.

+++ Examples/ScalarField/StateVariables.hpp
@@ -24,13 +24,13 @@ enum
 
 namespace StateVariables
 {
-static const std::array<std::string, NUM_VARS - NUM_CCZ4_VARS>
+static const std::array<std::string, static_cast<int>(NUM_VARS) - static_cast<int>(NUM_CCZ4_VARS)>
     additional_names = {"phi", "Pi"};
 
 static const std::array<std::string, NUM_VARS> names =
     ArrayTools::concatenate(CCZ4Variables::names, additional_names);
 
-static const std::array<BCParity, NUM_VARS - NUM_CCZ4_VARS>
+static const std::array<BCParity, static_cast<int>(NUM_VARS) - static_cast<int>(NUM_CCZ4_VARS)>
     additional_parities = {BCParity::even, BCParity::even};
 
 static const std::array<BCParity, NUM_VARS> parities =
